Fixed bug that the KnapsackSolver was not terminating
The Problem in the TestCase was too small for the given CoolingScheme
and Termination. The KnapsackSolver should pack each item only once in
the bag. The given capacity in the testcase was to high. It is still a
bad example, since the solver takes too many steps. You could bruteforce
this small problem in just a few steps.